+++ frontend/package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^18.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "rsuite": "^5.15.2",
     "typescript": "^4.7.4",
